Refactor: extract ToC into separate file and gate experimental logic
