+++ src/js/site/index.js
@@ -1,6 +1,7 @@
 import * as main from './main';
 import * as nav from './nav';
 import * as sidebar from './sidebar';
+import * as toc from './toc';
 import * as user from './user';
 
-export default { main, nav, sidebar, user };
+export default { main, nav, toc, sidebar, user };
